+++ test/mongo_driver_3/collection_test.clj
@@ -184,7 +184,7 @@
 
 ;;; Integration
 
-; docker run -it --rm -p 27017:27017 mongo:4.2
+; docker run -it --rm -p 27017:27017 mongo:4.2 --replset rs1
 
 (def client (atom nil))